Moved backup settings to GeneralTab
