+++ src/main/java/org/jabref/gui/preferences/file/FileTab.fxml
@@ -6,12 +6,8 @@
 <?import javafx.scene.control.RadioButton?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.ToggleGroup?>
-<?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.layout.GridPane?>
-<?import org.jabref.gui.icon.JabRefIconView?>
-
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.file.FileTab">
@@ -42,19 +38,6 @@
     </HBox>
 
     <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
-        <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <CheckBox fx:id="createBackup" text="%Create backup"></CheckBox>
-        <TextField fx:id="backupDirectory" HBox.hgrow="ALWAYS"/>
-        <Button fx:id="browseDirectory" onAction="#backupFileDirBrowse"
-                styleClass="icon-button,narrow"
-                prefHeight="20.0" prefWidth="20.0" GridPane.columnIndex="2">
-            <graphic>
-                <JabRefIconView glyph="OPEN"/>
-            </graphic>
-            <tooltip>
-                <Tooltip text="%Browse"/>
-            </tooltip>
-        </Button>
-    </HBox>
-    
+
+
 </fx:root>
