+++ src/main/java/org/jabref/gui/preferences/general/GeneralTab.fxml
@@ -1,12 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.control.Button?>
 <?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.Tooltip?>
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
+<?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
+<?import org.jabref.gui.icon.JabRefIconView?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.general.GeneralTab">
@@ -35,4 +40,19 @@
     <CheckBox fx:id="openLastStartup" text="%Open last edited libraries on startup"/>
     <CheckBox fx:id="showAdvancedHints"
               text="%Show advanced hints (i.e. helpful tooltips, suggestions and explanation)"/>
+
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <CheckBox fx:id="createBackup" text="%Create backup"/>
+        <TextField fx:id="backupDirectory" HBox.hgrow="ALWAYS"/>
+        <Button fx:id="browseDirectory" onAction="#backupFileDirBrowse"
+                styleClass="icon-button,narrow"
+                prefHeight="20.0" prefWidth="20.0" GridPane.columnIndex="2">
+            <graphic>
+                <JabRefIconView glyph="OPEN"/>
+            </graphic>
+            <tooltip>
+                <Tooltip text="%Browse"/>
+            </tooltip>
+        </Button>
+    </HBox>
 </fx:root>
