+++ src/main/java/org/jabref/gui/preferences/general/GeneralTabViewModel.java
@@ -1,5 +1,6 @@
 package org.jabref.gui.preferences.general;
 
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
@@ -10,15 +11,20 @@ import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.transformation.SortedList;
 
 import org.jabref.gui.DialogService;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
+import org.jabref.gui.util.DirectoryDialogConfiguration;
 import org.jabref.logic.l10n.Language;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseMode;
+import org.jabref.preferences.FilePreferences;
 import org.jabref.preferences.GeneralPreferences;
+import org.jabref.preferences.PreferencesService;
 import org.jabref.preferences.TelemetryPreferences;
 
 public class GeneralTabViewModel implements PreferenceTabViewModel {
@@ -34,17 +40,22 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
     private final BooleanProperty openLastStartupProperty = new SimpleBooleanProperty();
     private final BooleanProperty showAdvancedHintsProperty = new SimpleBooleanProperty();
 
+    private final BooleanProperty createBackupProperty = new SimpleBooleanProperty();
+    private final StringProperty backupDirectoryProperty = new SimpleStringProperty("");
+
     private final DialogService dialogService;
     private final GeneralPreferences generalPreferences;
     private final TelemetryPreferences telemetryPreferences;
+    private final FilePreferences filePreferences;
 
     private final List<String> restartWarning = new ArrayList<>();
 
     @SuppressWarnings("ReturnValueIgnored")
-    public GeneralTabViewModel(DialogService dialogService, GeneralPreferences generalPreferences, TelemetryPreferences telemetryPreferences) {
+    public GeneralTabViewModel(PreferencesService preferencesService, DialogService dialogService) {
         this.dialogService = dialogService;
-        this.generalPreferences = generalPreferences;
-        this.telemetryPreferences = telemetryPreferences;
+        this.generalPreferences = preferencesService.getGeneralPreferences();
+        this.telemetryPreferences = preferencesService.getTelemetryPreferences();
+        this.filePreferences = preferencesService.getFilePreferences();
     }
 
     public void setValues() {
@@ -60,6 +71,9 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
         collectTelemetryProperty.setValue(telemetryPreferences.shouldCollectTelemetry());
         openLastStartupProperty.setValue(generalPreferences.shouldOpenLastEdited());
         showAdvancedHintsProperty.setValue(generalPreferences.shouldShowAdvancedHints());
+
+        createBackupProperty.setValue(filePreferences.shouldCreateBackup());
+        backupDirectoryProperty.setValue(filePreferences.getBackupDirectory().toString());
     }
 
     public void storeSettings() {
@@ -84,6 +98,9 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
         generalPreferences.setShowAdvancedHints(showAdvancedHintsProperty.getValue());
 
         telemetryPreferences.setCollectTelemetry(collectTelemetryProperty.getValue());
+
+        filePreferences.createBackupProperty().setValue(createBackupProperty.getValue());
+        filePreferences.backupDirectoryProperty().setValue(Path.of(backupDirectoryProperty.getValue()));
     }
 
     @Override
@@ -131,5 +148,20 @@ public class GeneralTabViewModel implements PreferenceTabViewModel {
 
     public BooleanProperty showAdvancedHintsProperty() {
         return this.showAdvancedHintsProperty;
+    }
+
+    public BooleanProperty createBackupProperty() {
+        return this.createBackupProperty;
+    }
+
+    public StringProperty backupDirectoryProperty() {
+        return this.backupDirectoryProperty;
+    }
+
+    public void backupFileDirBrowse() {
+        DirectoryDialogConfiguration dirDialogConfiguration =
+                new DirectoryDialogConfiguration.Builder().withInitialDirectory(Path.of(backupDirectoryProperty().getValue())).build();
+        dialogService.showDirectorySelectionDialog(dirDialogConfiguration)
+                     .ifPresent(dir -> backupDirectoryProperty.setValue(dir.toString()));
     }
 }
